# handlers/handler_cities.go
package handlers

import (
	"github.com/gin-gonic/gin"
	"ynov_immo/models"
)

func init() {
	groupApi.GET("city", cityAll)
	groupApi.GET("city/:id", cityOne)
	groupApi.POST("city", cityCreate)
	groupApi.PATCH("city", cityUpdate)
	groupApi.DELETE("city/:id", cityDelete)
}

//All
func cityAll(c *gin.Context) {
	mdl := models.City{}
	query := &models.PaginationQuery{}
	err := c.ShouldBindQuery(query)
	if handleError(c, err) {
		return
	}
	list, total, err := mdl.All(query)
	if handleError(c, err) {
		return
	}
	jsonPagination(c, list, total, query)
}

//One
func cityOne(c *gin.Context) {
	var mdl models.City
	id, err := parseParamID(c)
	if handleError(c, err) {
		return
	}
	mdl.Id = id
	data, err := mdl.One()
	if handleError(c, err) {
		return
	}
	jsonData(c, data)
}

//Create
func cityCreate(c *gin.Context) {
	var mdl models.City
	err := c.ShouldBind(&mdl)
	if handleError(c, err) {
		return
	}
	err = mdl.Create()
	if handleError(c, err) {
		return
	}
	jsonData(c, mdl)
}

//Update
func cityUpdate(c *gin.Context) {
	var mdl models.City
	err := c.ShouldBind(&mdl)
	if handleError(c, err) {
		return
	}
	err = mdl.Update()
	if handleError(c, err) {
		return
	}
	jsonSuccess(c)
}

//Delete
func cityDelete(c *gin.Context) {
	var mdl models.City
	id, err := parseParamID(c)
	if handleError(c, err) {
		return
	}
	mdl.Id = id
	err = mdl.Delete()
	if handleError(c, err) {
		return
	}
	jsonSuccess(c)
}